adds some defensive scripting

+++ templify.ps1
@@ -1,33 +1,76 @@
 $targetFolder=$args[0]
 
-Write-Host "Setting up repository structure for " $targetFolder
+if (Test-Path $targetFolder) {
 
-# Constants
-$vsGitIgnore=".gitignore"
-$readme="README.md"
-$slnTemplate="template.sln"
-$reponame= Split-Path $targetFolder -leaf
+    Write-Host "Setting up repository structure for " $targetFolder
 
-
-# Script
-Copy-Item $vsGitIgnore -Destination $targetFolder
-Copy-Item $slnTemplate -Destination $targetFolder
-Rename-Item -path $targetFolder/$slnTemplate -newname $reponame-ci.sln
-New-Item $targetFolder/README.md -type file
-New-Item $targetFolder/drop-solution-file-here.txt -type file
+    # Constants
+    $vsGitIgnore=".gitignore"
+    $readme="README.md"
+    $slnTemplate="template.sln"
+    $reponame= Split-Path $targetFolder -leaf
 
 
-mkdir $targetFolder/src
-New-Item $targetFolder/src/drop-your-csproj-folders-here.txt -type file
+    # Script
+    if (Test-Path $targetFolder/$vsGitIgnore -eq $false) {
+        Copy-Item $vsGitIgnore -Destination $targetFolder        
+    }
 
-mkdir $targetFolder/build_automation
-New-Item $targetFolder/build_automation/drop-your-build-scripts-here.txt -type file
+    if (Test-Path $targetFolder/$reponame-ci.sln -eq $false) {
+        Copy-Item $slnTemplate -Destination $targetFolder
+        Rename-Item -path $targetFolder/$slnTemplate -newname $reponame-ci.sln
+    }
 
-mkdir $targetFolder/tools
-New-Item $targetFolder/tools/drop-tools-like-xunitrunner-here.txt -type file
+    if (Test-Path $targetFolder/README.md -eq $false) {
+        New-Item $targetFolder/README.md -type file
+    }
 
-mkdir $targetFolder/logs
-New-Item $targetFolder/logs/logfiles-go-here.txt -type file
+    if (Test-Path $targetFolder/drop-solution-file-here.txt -eq $true) {
+        Remove-Item $targetFolder/drop-solution-file-here.txt
+    }
+    
+    if (Test-Path $targetFolder/src) {
+        New-Item $targetFolder/src/drop-your-csproj-folders-here.txt -type file -force        
+    }
+    else {
+        mkdir $targetFolder/src
+        New-Item $targetFolder/src/drop-your-csproj-folders-here.txt -type file
+    }
+    
+    if (Test-Path $targetFolder/build_automation/) {
+        New-Item $targetFolder/build_automation/drop-your-build-scripts-here.txt -type file -force       
+    }
+    else {
+        mkdir $targetFolder/build_automation
+        New-Item $targetFolder/build_automation/drop-your-build-scripts-here.txt -type file
+    }
 
-mkdir $targetFolder/artifact
-New-Item $targetFolder/artifact/generated-NuGet-Packages-Go-here.txt -type file
+    
+    if (Test-Path $targetFolder/tools ) {
+        New-Item $targetFolder/tools/drop-tools-like-xunitrunner-here.txt -type file -force        
+    }
+    else {
+        mkdir $targetFolder/tools                
+        New-Item $targetFolder/tools/drop-tools-like-xunitrunner-here.txt -type file
+    }
+
+    if (Test-Path $targetFolder/logs) {
+        New-Item $targetFolder/logs/logfiles-go-here.txt -type file -force             
+    }
+    else {
+        mkdir $targetFolder/logs
+        New-Item $targetFolder/logs/logfiles-go-here.txt -type file -force 
+    }
+    
+
+    if (Test-Path $targetFolder/artifact) {
+        New-Item $targetFolder/artifact/generated-NuGet-Packages-Go-here.txt -type file -force         
+    }
+    else {
+        mkdir $targetFolder/artifact
+        New-Item $targetFolder/artifact/generated-NuGet-Packages-Go-here.txt -type file -force 
+    }
+    
+}
+
+
